fedi: structures v2
these structures are written in such a way to allow multiple types in
one spot, as per the activitypub spec.
i have no idea on the effects it will have for federating with fchannel,
and it is slower than the previous implementation, but it is much easier
to work with.