fedi: force first char of apid to be a letter
this is an awful fix, but hey, it works i guess
this is done because on some endpoints that take in a post ID, it'll
also accept the activitypub id to remain compatable with fchannel.
it'll attempt to parse it in decimal, and if that fails, it'll fall back
to hexadecimal.

so what happens when you get an id like 12345678? valid hex, valid
decimal. you see the problem?
i could've just fixed how it's done on those endpoints, by simply
parsing it as hex if the decimal lookup fails but eh. later.