meta: DONATE.md
this is intentionally kept away from the actual code to not mislead
users into thinking they're donating to the instance admin when actually
all of their money goes to me.
instance admins, if you wish to accept donations, the best way to do
this is to just edit views/index.html or views/faq.html and to put it
there.