sqlite3: database upgrades, no intervention needed
previously i said that you would need to do this yourself, i decided
against this and detect if it should be initialized or not.
existing databases will simply be upgraded to the latest version of the
schema automatically.

however, i still suggest taking a backup before doing anything in case
this blows up in your face, in which case please file a bug report and
call me stupid